requirements.txt
asttokens==3.0.0
cmake==3.31.4
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
executing==2.2.0
filelock==3.17.0
fonttools==4.55.8
fsspec==2024.12.0
icecream==2.1.4
imageio==2.37.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pillow_heif==0.21.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.1.10
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
utils==1.0.2
wheel==0.45.1
